Apply middleware in post create, update, delete

+++ server/controllers/post.js
@@ -5,6 +5,7 @@ import Posts from '../models/post.js';
 export const createPost = async (req, res) => {
     const postInfo = req.body;
     try {
+        if (!req.userId) return res.json({message: "Unathenticated"});
         const newPost = new Posts(postInfo);
 
         await newPost.save();
@@ -27,6 +28,7 @@ export const getPosts = async (req, res) => {
 export const updatePost = async (req, res) => {
     const { _id } = req.params;
     const post = req.body;
+    if (!req.userId) return res.json({message: "Unathenticated"});
 
     if (!mongoose.Types.ObjectId.isValid(_id)) return res.status(404).send("No post with that _id");
 
@@ -37,7 +39,8 @@ export const updatePost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
     const { _id } = req.params;
-
+    if (!req.userId) return res.json({message: "Unathenticated"});
+    
     if (!mongoose.Types.ObjectId.isValid(_id)) return res.status(404).send("No post with that id");
 
     await Posts.findByIdAndRemove(_id);
